Make use of APR gethostbyname equivalent when APR is available. Remove AS/400 specific code and fix some short to int in port. We'll use integer everywhere and cast to short only in jk_resolve
git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@298789 13f79535-47bb-0310-9956-ffa450edef68

+++ jk/native/common/jk_ajp12_worker.c
@@ -213,7 +213,7 @@ static int JK_METHOD validate(jk_worker_t *pThis,
                p->name, host, port);
 
         if(port > 1024 && host) {
-            if(jk_resolve(host, (short)port, &p->worker_inet_addr)) {
+            if(jk_resolve(host, port, &p->worker_inet_addr)) {
                 return JK_TRUE;
             }
             jk_log(l, JK_LOG_ERROR, "In jk_worker_t::validate, resolve failed\n");
