+++ jk/native/common/jk_ajp_common.c
@@ -1366,7 +1366,7 @@ int ajp_validate(jk_worker_t *pThis,
                p->name, host, port);
 
         if(port > 1024 && host) {
-            if(jk_resolve(host, (short)port, &p->worker_inet_addr)) {
+            if(jk_resolve(host, port, &p->worker_inet_addr)) {
                 return JK_TRUE;
             }
             jk_log(l, JK_LOG_ERROR,
